Remove binary_neuron_impl.h
Since the _impl.h file provides required definitions of template
functions, it always needs to be included along with the .h
file. To avoid having two seperate files for the definition of the
binary_neuron template, all functions are moved to the .h file.
